pyproject.toml
[project]
name = "eng-lakehouse-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "faker>=40.1.2",
    "psycopg2>=2.9.11",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.46",
    "tqdm>=4.67.2",
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
]
